Lowercase drive letters when getting paths from cargo check
